Move sub-classes of "TabPage" and "ToolTip" classes to their own files
