+++ src/System.Windows.Forms/src/System/Windows/Forms/TabPage.cs
@@ -22,7 +22,7 @@ namespace System.Windows.Forms
     [DesignTimeVisible(false)]
     [DefaultEvent("Click")]
     [DefaultProperty("Text")]
-    public class TabPage : Panel
+    public partial class TabPage : Panel
     {
         private ImageList.Indexer _imageIndexer;
         private string _toolTipText = string.Empty;
@@ -725,36 +725,6 @@ namespace System.Windows.Forms
             if (ParentInternal is TabControl parent)
             {
                 parent.UpdateTab(this);
-            }
-        }
-
-        /// <summary>
-        ///  Our control collection will throw an exception if you try to add other tab pages.
-        /// </summary>
-        public class TabPageControlCollection : ControlCollection
-        {
-            /// <summary>
-            ///  Creates a new TabPageControlCollection.
-            /// </summary>
-            public TabPageControlCollection(TabPage owner) : base(owner)
-            {
-            }
-
-            /// <summary>
-            ///  Adds a child control to this control. The control becomes the last control
-            ///  in the child control list. If the control is already a child of another
-            ///  control it is first removed from that control. The tab page overrides
-            ///  this method to ensure that child tab pages are not added to it, as these
-            ///  are illegal.
-            /// </summary>
-            public override void Add(Control value)
-            {
-                if (value is TabPage)
-                {
-                    throw new ArgumentException(SR.TabControlTabPageOnTabPage);
-                }
-
-                base.Add(value);
             }
         }
     }
